Allow an arbitrary order of many components in the PDDL domain specification
Previously, there were only cryptic error messages, when a user specified the
components in the PDDL domain specification in a wrong order (e.g. predicates
before constants). Since adding better error messages to the current state of
the code makes it even more ugly, for the moment we simply allow an arbitrary
order of the requirement, types, constants, predicates and functions
specification. The PDDL BNF is more strict on this, so we might want to change
this again in a future refactoring.

--HG--
branch : issue163